% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{analyze.preformance}
\alias{analyze.preformance}
\title{Analyzes performance of function}
\usage{
analyze.preformance(f, file.name = "tmp.out")
}
\arguments{
\item{f}{function which will be evaluated}

\item{file.name}{temporary file, where the output will be stored}
}
\description{
Analyzes performance of specified function. Usefull for performance issues.
}
\author{
Filip Dechterenko
}
