% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{cart2pol}
\alias{cart2pol}
\title{Convert cartesian coordinates to polar}
\usage{
cart2pol(x, y, deg = T)
}
\arguments{
\item{x}{coordinate x}

\item{y}{coordinate y}

\item{deg}{boolean, if output should be in degrees instead of radians}
}
\description{
Convert cartesian coordinates to polar
}
\author{
Filip Dechterenko
}
