% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{changescale}
\alias{changescale}
\title{Changes scale of the variable}
\usage{
changescale(x, oldmin, oldmax, newmin, newmax)
}
\arguments{
\item{x}{vector or matrix with values}

\item{oldmin}{old minimum}

\item{oldmax}{old maximum}

\item{newmin}{new minimum}

\item{newmax}{new maximum}
}
\description{
Changes scale of the variable from one range to another
}
\author{
Filip Dechterenko
}
