% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{cohen.d}
\alias{cohen.d}
\title{Computes cohen's d}
\usage{
cohen.d(x, y)
}
\arguments{
\item{x}{vector of values}

\item{y}{vector of values}
}
\description{
Computes cohen's d for two vectors (pooled version)
}
\author{
Filip Dechterenko
}
