% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT.R
\name{compute_bias}
\alias{compute_bias}
\title{Computes d' from hits and false alarms}
\usage{
compute_bias(H, FA, type = "C", should_correct = F, max_bias = 4)
}
\arguments{
\item{H}{hit rate (pecentages)}

\item{FA}{false alarm rate (pecentages)}

\item{type}{type of bias estimate, possible values are "C", C_rel","lnB". Default value is C}

\item{should_correct}{boolean, whether we should correct perfect performance}

\item{max_bias}{to which values we should correct perfect performance}
}
\value{

}
\description{
Bias correction currently works only for C estimate of bias
}
\examples{
compute_bias(.4,.8)
compute_bias(.2,.7, type = "lnB")
compute_bias(.4, 1, should_correct = T)
}
