% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT.R
\name{compute_dprime}
\alias{compute_dprime}
\title{Computes d' from hits and false alarms}
\usage{
compute_dprime(H, FA, should_correct = F, max_d = 4)
}
\arguments{
\item{H}{hit rate (pecentages)}

\item{FA}{false alarm rate (pecentages)}

\item{should_correct}{boolean, whether we should correct perfect performance}

\item{max_d}{to which values we should correct perfect performance}
}
\value{

}
\description{
Uses traditional formula
}
\examples{
compute_dprime(.4,.8)
compute_dprime(.4, 1, should_correct = T)
}
