% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT.R
\name{compute_FA_from_dprime}
\alias{compute_FA_from_dprime}
\title{Computes sensitivity d' from hit rate}
\usage{
compute_FA_from_dprime(d, H)
}
\arguments{
\item{d}{sensitivity}

\item{H}{hit rate}
}
\value{

}
\description{
This formula assumes zero bias
}
\examples{
compute_FA_from_dprime(2,.8)
}
