% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{create.time.measure}
\alias{create.time.measure}
\title{Create time measure object}
\usage{
create.time.measure(n)
}
\arguments{
\item{n}{number of cases which have to evaluated}
}
\description{
Create and initialize time measure object. This object can be used for printing remaining time in some simulations
}
\author{
Filip Dechterenko
}
