% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_lrt}
\alias{format_lrt}
\title{Formats output from likelihood ratio test when comparing two lmer models}
\usage{
format_lrt(av)
}
\arguments{
\item{av}{results of anova(lm1,lm2) function}
}
\value{

}
\description{
Formats output from likelihood ratio test when comparing two lmer models
}
