% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{pol2cart}
\alias{pol2cart}
\title{Convert polar coordinates to cartesian}
\usage{
pol2cart(size, angle, deg = T)
}
\arguments{
\item{size}{distance from the origin}

\item{angle}{anlge parameter of the point}

\item{deg}{boolean, if output should be in degrees instead of radians}
}
\description{
Convert polar coordinates to cartesian
}
\author{
Filip Dechterenko
}
