% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{prettyprint}
\alias{prettyprint}
\title{Pretty prints values for publication}
\usage{
prettyprint(x, p = 2)
}
\arguments{
\item{x}{vector or matrix with values}

\item{p}{how many decimal places should print precisely}
}
\description{
Pretty prints values for publication. Values are ounded to specified precision with trailing zeros.
}
\author{
Filip Dechterenko
}
