% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{test_dir}
\alias{test_dir}
\title{Tests whether directory exist}
\usage{
test_dir(pth, should_exist = F)
}
\arguments{
\item{pth}{path to directory}

\item{should_exist}{whether the directory should exist.}
}
\value{

}
\description{
this function can also create it, if it is specified
}
