% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{update.tm}
\alias{update.tm}
\title{Update estimation of remaining time.}
\usage{
\method{update}{tm}(object, ...)
}
\arguments{
\item{object}{initialized time measure object}

\item{...}{this is not used, only for compatibilty issues}
}
\description{
Update estimation of remaining time. We should call this method after one item was computed.
}
\author{
Filip Dechterenko
}
