% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{print.tm}
\alias{print.tm}
\title{Prints remaining time of computation}
\usage{
\method{print}{tm}(x)
}
\arguments{
\item{x}{initialized time measure object}
}
\description{
Prints remaining time of computation. First \code{update.tm} should be called
}
\author{
Filip Dechterenko
}
